pyproject.toml
[project]
name = "dbx-airflow"
version = "0.1.0"
description = "testing and learning Databricks + Airflow"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-databricks>=6.10.0",
    "apache-airflow>=2.10.2",
    "databricks-connect==15.1.*",
    "databricks-sdk>=0.33.0",
]
